correct casting in sprintf's to shutup compiler warning
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@390 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Base/GemPixDualObj.cpp
@@ -158,7 +158,7 @@ void GemPixDualObj :: processDualImage(imageStruct &left, imageStruct &right){
     lformat ="YUV";break;
   default:
     lformat = new char[6];
-    sprintf(lformat,"0x%04X", left.format);
+    sprintf(lformat,"0x%04X", (unsigned int)left.format);
   }
   switch (right.format) {
   case GL_RGBA:
@@ -170,7 +170,7 @@ void GemPixDualObj :: processDualImage(imageStruct &left, imageStruct &right){
     rformat ="YUV";break;
   default:
     rformat = new char[6];
-    sprintf(rformat, "0x%04X", left.format);
+    sprintf(rformat, "0x%04X", (unsigned int)left.format);
   }
   
   error("processDualImage: no method to combine (%s) and (%s)",
